# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API do Dicionário",
        "description": "Documentação da API do Dicionário. Fullstack Challenge 🏅 - Dictionary",
        "contact": {
            "email": "suporte@dicionario.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/entries/en/{id_user}/{word}/favorite": {
            "post": {
                "tags": [
                    "Gerenciamento de Palavras"
                ],
                "summary": "Adiciona a lista de favoritos a palavra selecionada pelo o usuário",
                "operationId": "9c200ff8e6cb932af2291816c11af00c",
                "responses": {
                    "200": {
                        "description": "Added to favorites successfully!"
                    },
                    "500": {
                        "description": "An error occurred, try again!"
                    },
                    "400": {
                        "description": "Undefined Word!, The word has already been added to the favorites list!"
                    },
                    "403": {
                        "description": "Unauthorized access."
                    }
                }
            }
        },
...
        "/api/entries/en/{id_user}/{word}/unfavorite": {
            "delete": {
                "tags": [
                    "Gerenciamento de Palavras"
                ],
                "summary": "Realiza a exclusão das palavras adicionadas na lista de favoritos do usuário.",
                "operationId": "b5ae24b55478e679c1f8b63556de38f4",
                "responses": {
                    "204": {
                        "description": "Removed from favorites successfully!"
                    },
                    "500": {
                        "description": "An error occurred, try again!"
                    },
                    "400": {
                        "description": "The word is not included in the favorites list!, Undefined Word!"
                    },
                    "403": {
                        "description": "Unauthorized access."
                    }
                }
            }
        },
        "/api/entries/en/{word}": {
            "post": {
                "tags": [
                    "Gerenciamento de Palavras"
                ],
                "summary": "Retorna os dados da palavra pesquisada pelo o usuário.",
                "operationId": "fc54ab1373b4bf0e7d54a566f5ba4ef1",
                "responses": {
                    "500": {
                        "description": "An error occurred, try again!"
                    },
                    "400": {
                        "description": "No Definitions Found. Sorry pal, we couldn`t find definitions for the word you were looking for! You can try the search again at later time or head to the web instead."
                    }
                }
            }
        },
...
    },
...
}